rename controller pkg/registry/core/controller to pkg/registry/core/replicationcontroller
