+++ pkg/registry/core/replicationcontroller/BUILD
