+++ pkg/registry/core/replicationcontroller/strategy_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package replicationcontroller
 
 import (
 	"strings"
